Move function from data.py to recording.py
Move get_recording_cluster_id_using_recording_mbid from data.py
to recording.py as it used by recording.py only and makes
more sense to have it in recording.py.
