+++ messybrainz/db/recording.py
@@ -1,5 +1,4 @@
 from messybrainz import db
-from messybrainz.data import get_recording_cluster_id_using_recording_mbid
 from sqlalchemy import text
 
 
@@ -99,6 +98,29 @@ def truncate_tables():
         connection.execute(query)
         query = text("TRUNCATE TABLE recording_redirect")
         connection.execute(query)
+
+
+def get_recording_cluster_id_using_recording_mbid(connection, recording_mbid):
+    """Returns cluster_id for a required recording MBID.
+
+    Args:
+        connection: the sqlalchemy db connection to be used to execute queries
+        recording_mbid (UUID): recording MBID for the cluster.
+
+    Returns:
+        cluster_id (UUID): MSID that represents the cluster if it exists else None.
+    """
+
+    query = text("""SELECT recording_cluster_id
+                      FROM recording_redirect
+                     WHERE recording_mbid = :recording_mbid""")
+
+    cluster_id = connection.execute(query, {"recording_mbid": recording_mbid})
+
+    if cluster_id.rowcount:
+        return cluster_id.fetchone()['recording_cluster_id']
+    else:
+        return None
 
 
 def create_recording_clusters():
